distributed_kv_store/requirements.txt
django
djangorestframework
requests
psycopg2
